refactor(): domain 패키지 별로 묶음
