+++ src/main/java/lotto/domain/lotto/LottoCount.java
@@ -1,4 +1,4 @@
-package lotto.domain;
+package lotto.domain.lotto;
 
 import lotto.exception.IllegalLottoCountException;
