+++ src/main/java/lotto/domain/lotto/LottoTicket.java
@@ -1,4 +1,4 @@
-package lotto.domain;
+package lotto.domain.lotto;
 
 import lotto.exception.IllegalLottoNumbersException;
