+++ src/main/java/lotto/domain/result/LottoResult.java
@@ -1,4 +1,6 @@
-package lotto.domain;
+package lotto.domain.result;
+
+import lotto.domain.Money;
 
 import java.util.ArrayList;
 import java.util.Collections;
