+++ src/main/java/lotto/domain/result/Prize.java
@@ -1,4 +1,6 @@
-package lotto.domain;
+package lotto.domain.result;
+
+import lotto.domain.Money;
 
 import java.util.Arrays;
