+++ src/test/java/lotto/domain/LottoCountTest.java
@@ -1,5 +1,6 @@
 package lotto.domain;
 
+import lotto.domain.lotto.LottoCount;
 import lotto.exception.IllegalLottoCountException;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.params.ParameterizedTest;
