+++ src/test/java/lotto/domain/LottoResultTest.java
@@ -1,5 +1,7 @@
 package lotto.domain;
 
+import lotto.domain.result.LottoResult;
+import lotto.domain.result.Prize;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
