+++ src/test/java/lotto/domain/LottoTicketTest.java
@@ -1,5 +1,7 @@
 package lotto.domain;
 
+import lotto.domain.lotto.LottoNumber;
+import lotto.domain.lotto.LottoTicket;
 import lotto.exception.IllegalLottoNumbersException;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
